统一采用构造器注入加Lazy标记解决循环依赖

+++ src/main/java/com/example/vnollxonlinejudge/utils/GetScore.java
@@ -1,7 +1,7 @@
 package com.example.vnollxonlinejudge.utils;
 
 public class GetScore {
-    public static Long calculateScore(int passCount, int penaltyTime) {
-        return ((long) (-passCount) << 32) | (penaltyTime & 0xFFFFFFFFL);
+    public static Long calculateScore(Long passCount, Long penaltyTime) {
+        return ( (-passCount) << 32) | (penaltyTime & 0xFFFFFFFFL);
     }
 }
